+++ src/three/TilesGroup.d.ts
@@ -1,10 +1,11 @@
-import { Group } from 'three';
+import { Group, Matrix4 } from 'three';
 import { TilesRenderer } from './TilesRenderer';
 
 export class TilesGroup extends Group {
 
 	readonly isTilesGroup: true;
-	tilesRenderer : TilesRenderer;
+	tilesRenderer: TilesRenderer;
+	matrixWorldInverse: Matrix4;
 	constructor( tilesRenderer : TilesRenderer );
 
 }